Initial commit (read below)
Initial commit of this repository containing the entire contents of my
UBELIX HOME directory (/home/ubelix/sm/dw07r324).

~/jobs/2015-05-16 contains a setup for running all the tests (complement
internal/external GOAL/Michel, testing completeness, and testing uni-
versality). This setup consists of the directory structure, and run
scripts, ready to be executed or sourced.

The setup includes the execution of 'rank -tr -ro' on 10 partitions of
the GOAL test set (because it can't finish the whole test set in 72 hours).

Note that on UBELIX, the remote for the corresponding GitHub repository
must be added as SSH rather than HTTPS. Pushing over HTTPS does not seem
to work from UBELIX. To do so, a RSA key pair has been generated on UBELIX,
and the public key has been put on GitHub.